requirements.txt
cffi==1.11.5
pycparser==2.19
pycryptodome==3.7.2
PyNaCl==1.3.0
six==1.12.0
